pyproject.toml
[tool.poetry]
name = "maven-ai-engineer"
version = "0.1.0"
description = ""
authors = ["Faris <frs@hbb.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.1.1"
openai = "^1.25.1"
python-dotenv = "^1.0.1"
black = "^24.4.2"
pandas = "^2.2.2"
tiktoken = "^0.6.0"
langchain = "^0.1.17"
numpy = "^1.26.4"
scipy = "^1.13.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
